Initial fork
Changed logging features to use the devnull logger library (which now works) and a
bunch of other stuff (which doesn't).

The biggest problem at the moment is that the camera keeps on sending the same pictures
over and over again, probably since the app can't tell that the upload is already on
the server. This may be due to the fact that i can't respond with a correct `offset`
field in the `GetPhotoStatus` SOAP request. I might be able to figure out some details
by "wiresharking" it with my proprietary-OS laptop.

I am not a coder (and specifically not a Node.js coder) so please pardon my incompetence.

This version is supposed to be running on a Raspberry Pi. The Pi this is running on
also works as a wireless access point.